+++ documentation/changelog.md
@@ -18,6 +18,8 @@ description: List of all the changes and current versions.
 + Add Additional Interaction Key Press Cancels
 + Fix Vehicle Door Synchronization
 + Added /seatbelt, /sb
++ Removed Automatic Respawn
++ Added /acceptdeath
 ```
 
 ## 1.5.0 - January 28-31, 2021